requirements.txt
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
paramiko>=3.4,<4.0
selenium>=4.20,<5.0
webdriver-manager>=4.0,<5.0
